Fixed Login/Register Issue
There was an issue where if you click on register button from login page and then click the go back button, it creates a new login page on top of current register page. Instead I made it pop the register page taking you back to the old login page. This is better so we will not have many pages on top of eachother